qinr/requirements.txt
appdirs==1.4.4
autograd==1.5
autoray==0.6.7
backcall==0.2.0
cachetools==5.3.2
catboost==1.0.4
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.12
cloudpickle==2.0.0
cryptography==41.0.5
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
dill==0.3.7
entrypoints==0.4
fonttools==4.31.1
future==0.18.3
graphviz==0.19.1
grpcio==1.44.0
h5py==3.6.0
ibm-cloud-sdk-core==3.17.2
ibm-platform-services==0.45.1
idna==3.3
imageio==2.31.6
ipykernel==6.9.2
ipython==7.32.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.0
lazy_loader==0.3
lightgbm==3.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mpmath==1.3.0
nest-asyncio==1.5.4
networkx==3.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
parso==0.8.3
pbr==5.11.1
PennyLane==0.30.0
PennyLane-Lightning==0.31.0
PennyLane-qiskit==0.33.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
ply==3.11
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.8.0
pyparsing==3.0.7
pyspnego==0.10.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==22.3.0
qiskit==0.44.3
qiskit-aer==0.13.0
qiskit-ibm-provider==0.7.2
qiskit-ibm-runtime==0.14.0
qiskit-terra==0.25.3
requests==2.31.0
requests-ntlm==1.2.0
rustworkx==0.13.2
scikit-image==0.21.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
semantic-version==2.10.0
six==1.16.0
sklearn==0.0
spyder-kernels==2.2.1
stevedore==5.1.0
symengine==0.9.2
sympy==1.12
tenacity==8.0.1
threadpoolctl==3.1.0
tifffile==2023.7.10
toml==0.10.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.1
traitlets==5.1.1
typing_extensions==4.8.0
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==1.6.4
websockets==12.0
wurlitzer==3.0.2
xgboost==1.5.2
